Split GUI's tabs into into separate classes (#552)
* Schema tab

* Analytics tab

* optional tab

* Live tab

* About tab

* Output tab
