+++ src/daf_gui/tabs/__init__.py
@@ -1,0 +1,6 @@
+from .analytics import *
+from .optional import *
+from .schema import *
+from .about import *
+from .debug import *
+from .live import *
